code/requirements.txt
asgiref==3.7.2
av==10.0.0
backports.zoneinfo==0.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
coloredlogs==15.0.1
ctranslate2==3.18.0
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
faster-whisper==0.7.1
filelock==3.12.3
flatbuffers==23.5.26
fsspec==2023.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
Markdown==3.4.4
mpmath==1.3.0
numpy==1.24.4
onnxruntime==1.15.1
packaging==23.1
protobuf==4.24.2
pydub==0.25.1
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
sympy==1.12
tokenizers==0.13.3
tqdm==4.66.1
typing-extensions==4.7.1
urllib3==2.0.4
zipp==3.16.2
